requirements.txt
channels
django-rest-framework
